moved translations
