+++ packages/shared/src/locales/followers/followers.en.ts
