+++ packages/shared/src/locales/followers/followers.ru.ts
